backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
pymilvus==2.3.6

# Cache & Queue
redis==5.0.1

# AI & LLM
openai==1.10.0
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# CORS
fastapi-cors==0.0.6

pydantic[email]